Change RestResponse to Response and move to another package
