+++ anymarket-sdk-product/src/main/java/br/com/anymarket/draft/sdk/product/ProductFeedClient.java
@@ -5,7 +5,7 @@ import br.com.anymarket.draft.sdk.feed.Feed;
 import br.com.anymarket.draft.sdk.feed.MarkFeedAsReadRequest;
 import br.com.anymarket.draft.sdk.feed.MarkFeedAsReadResponse;
 import br.com.anymarket.sdk.http.HttpService;
-import br.com.anymarket.sdk.http.restdsl.RestResponse;
+import br.com.anymarket.sdk.http.Response;
 import com.fasterxml.jackson.core.type.TypeReference;
 
 import java.util.List;
@@ -17,14 +17,14 @@ public class ProductFeedClient extends HttpService {
     }
 
     public List<Feed> list() {
-        RestResponse response = execute(get("/products/feeds"));
+        Response response = execute(get("/products/feeds"));
         return response.to(new TypeReference<List<Feed>>() {
         });
     }
 
     public MarkFeedAsReadResponse markAsRead(Feed feed) {
         MarkFeedAsReadRequest request = new MarkFeedAsReadRequest(feed.getToken());
-        RestResponse response = executeWithBody(request, put("/products/feeds/" + feed.getId()));
+        Response response = executeWithBody(request, put("/products/feeds/" + feed.getId()));
         return response.to(MarkFeedAsReadResponse.class);
     }
 }
