+++ anymarket-sdk-stock/src/main/java/br/com/anymarket/sdk/stock/StockService.java
@@ -2,8 +2,8 @@ package br.com.anymarket.sdk.stock;
 
 import br.com.anymarket.sdk.SDKConstants;
 import br.com.anymarket.sdk.http.HttpService;
+import br.com.anymarket.sdk.http.Response;
 import br.com.anymarket.sdk.http.headers.IntegrationHeader;
-import br.com.anymarket.sdk.http.restdsl.RestResponse;
 import br.com.anymarket.sdk.stock.dto.Stock;
 import br.com.anymarket.sdk.stock.dto.StockCollection;
 import com.google.common.collect.Lists;
@@ -21,17 +21,17 @@ public class StockService extends HttpService {
             SDKConstants.ANYMARKET_HOMOLOG_API_ENDPOINT;
     }
 
-    public RestResponse updateStock(Stock stock, IntegrationHeader... headers) {
+    public Response updateStock(Stock stock, IntegrationHeader... headers) {
         return updateStock(Lists.newArrayList(stock), headers);
     }
 
-    public RestResponse updateStock(List<Stock> stocks, IntegrationHeader... headers) {
+    public Response updateStock(List<Stock> stocks, IntegrationHeader... headers) {
         StockCollection stockCollection = new StockCollection();
         stockCollection.addAllStock(stocks);
         return updateStock(stockCollection, headers);
     }
 
-    public RestResponse updateStock(StockCollection stocks, IntegrationHeader... headers) {
+    public Response updateStock(StockCollection stocks, IntegrationHeader... headers) {
         return executeWithBody(stocks.getStocks(),
             put(apiEndPoint + stocks.getPathURI(), headers)
         );
